env_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohttp=3.8.5=pypi_0
aiosignal=1.3.1=pypi_0
appdirs=1.4.4=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
black=23.9.1=pypi_0
blas=1.0=mkl
boltons=23.0.0=py310h06a4308_0
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.0=h5eee18b_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.7.22=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
cmake=3.27.5=pypi_0
coloredlogs=15.0.1=pypi_0
conda=23.7.4=py310h06a4308_0
conda-content-trust=0.2.0=py310h06a4308_0
conda-libmamba-solver=23.7.0=py310h06a4308_0
conda-package-handling=1.9.0=py310h5eee18b_1
cryptography=41.0.3=py310hdda0065_0
datasets=2.14.5=pypi_0
dill=0.3.7=pypi_0
docker-pycreds=0.4.0=pypi_0
filelock=3.12.4=pypi_0
flatbuffers=23.5.26=pypi_0
fmt=9.1.0=hdb19cb5_0
frozenlist=1.4.0=pypi_0
fsspec=2023.6.0=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.36=pypi_0
huggingface-hub=0.17.2=pypi_0
humanfriendly=10.0=pypi_0
icu=58.2=he6710b0_3
idna=3.4=py310h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46305
isort=5.12.0=pypi_0
jinja2=3.1.2=pypi_0
jsonpatch=1.32=pyhd3eb1b0_0
jsonpointer=2.1=pyhd3eb1b0_0
krb5=1.20.1=h143b758_1
ld_impl_linux-64=2.38=h1181459_1
libarchive=3.6.2=h6ac8c49_2
libcurl=8.1.1=h251f7ec_1
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libmamba=1.4.1=h2dafd23_1
libmambapy=1.4.1=py310h2dafd23_1
libnghttp2=1.52.0=h2d74bed_1
libsolv=0.7.22=he621ea3_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxml2=2.10.3=hcbfbd50_0
lit=16.0.6=pypi_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.3=pypi_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.6=py310h1128e8f_1
mkl_random=1.2.2=py310h1128e8f_1
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.15=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
numpy=1.25.2=py310h5f9d8c6_0
numpy-base=1.25.2=py310hb5e798b_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cublas-cu12=12.2.5.6=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-nvrtc-cu12=12.2.140=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu12=12.2.140=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cudnn-cu12=8.9.4.25=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
onnx=1.14.1=pypi_0
onnxruntime-gpu=1.15.1=pypi_0
onnxscript=0.1.0.dev20230918=pypi_0
openssl=3.0.10=h7f8727e_2
packaging=23.1=py310h06a4308_0
pandas=2.1.0=pypi_0
pathspec=0.11.2=pypi_0
pathtools=0.1.2=pypi_0
pcre2=10.37=he7ceb23_1
pip=23.2.1=py310h06a4308_0
platformdirs=3.10.0=pypi_0
pluggy=1.0.0=py310h06a4308_1
protobuf=4.24.3=pypi_0
psutil=5.9.5=pypi_0
pyarrow=13.0.0=pypi_0
pybind11-abi=4=hd3eb1b0_1
pycosat=0.6.4=py310h5eee18b_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py310h06a4308_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.13=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
reproc=14.2.4=h295c915_1
reproc-cpp=14.2.4=h295c915_1
requests=2.31.0=py310h06a4308_0
ruamel.yaml=0.17.21=py310h5eee18b_0
ruamel.yaml.clib=0.2.6=py310h5eee18b_1
sentry-sdk=1.31.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=68.0.0=py310h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.1=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorrt=8.6.1.post1=pypi_0
tensorrt-bindings=8.6.1=pypi_0
tensorrt-libs=8.6.1=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
toolz=0.12.0=py310h06a4308_0
torch=2.0.1=pypi_0
tqdm=4.65.0=py310h2f386ee_0
triton=2.0.0=pypi_0
typing-extensions=4.8.0=pypi_0
tzdata=2023.3=pypi_0
urllib3=1.26.16=py310h06a4308_0
wandb=0.15.10=pypi_0
wheel=0.38.4=py310h06a4308_0
xxhash=3.3.0=pypi_0
xz=5.4.2=h5eee18b_0
yaml-cpp=0.7.0=h295c915_1
yarl=1.9.2=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
